Updated Main Scene
Added Sphere game object to Hierarchy and renamed it to Planet.
Added Planet script component to Planet game object.
Added Planet material to Planet game object.
Assigned global variables to Planet script.
Created prefab of Planet game object.
Created an Empty GameObject in the Hierarchy and renamed it to Spawner.
Added the Spawner script to the Spawner game object.
Assigned references to Spawner script.
Created Empty GameObject in the Hierarchy and renamed it to Planets.